fix: change tickets.price column type from INT to DECIMAL(10,2)
The calculatePrice() method returns distance * 0.5 which is a float,
but the column was defined as INT causing silent precision loss.
DECIMAL(10,2) stores monetary values correctly (e.g. 150.50).

+++ database/migrations/create_table_tickets.php
@@ -10,7 +10,7 @@ class CreateTicketsTable
             passenger INT NOT NULL,
             route INT NOT NULL,
             vehicle INT NOT NULL,
-            price INT NOT NULL,
+            price DECIMAL(10,2) NOT NULL,
             date DATE NOT NULL,
             status TINYINT(1) DEFAULT 1,
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
